// frontend/StockManager/generated-sources/openapi/models/UserLoginDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * StockManager API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UserLoginDto
 */
export interface UserLoginDto {
    /**
     * 
     * @type {string}
     * @memberof UserLoginDto
     */
    email: string;
    /**
     * 
     * @type {string}
     * @memberof UserLoginDto
     */
    password: string;
}

/**
 * Check if a given object implements the UserLoginDto interface.
 */
export function instanceOfUserLoginDto(value: object): value is UserLoginDto {
    if (!('email' in value) || value['email'] === undefined) return false;
    if (!('password' in value) || value['password'] === undefined) return false;
    return true;
}

export function UserLoginDtoFromJSON(json: any): UserLoginDto {
    return UserLoginDtoFromJSONTyped(json, false);
}

export function UserLoginDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserLoginDto {
    if (json == null) {
        return json;
    }
    return {
        
        'email': json['email'],
        'password': json['password'],
    };
}

export function UserLoginDtoToJSON(json: any): UserLoginDto {
    return UserLoginDtoToJSONTyped(json, false);
}

export function UserLoginDtoToJSONTyped(value?: UserLoginDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'email': value['email'],
        'password': value['password'],
    };
}